Agregamos una exeption cuando no encuentra el job

+++ app/Services/JobService.php
@@ -10,6 +10,9 @@ class JobService
     public function deleteJob($id) {
         //Find the job with id x
         $job = Job::find($id);
+        if(!$job) {
+            throw new \Exception();
+        }
         //Delete physically from DB
         $job->delete();
     }
